+++ test/dart_ddi_test.dart
@@ -1,5 +1,3 @@
-import 'package:dart_ddi/dart_ddi.dart';
-
 import 'beans_test/add_decoratos_test.dart';
 import 'beans_test/application_future_test.dart';
 import 'beans_test/application_test.dart';
@@ -34,7 +32,6 @@ import 'event_test/timer_events_test.dart';
 import 'stream_test/stream_test.dart';
 
 void main() {
-  ddi.setDebugMode(false);
   //Basic Tests, with consists in register, get, dispose, remove
   singleton();
   application();